# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FounderAI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type SessionAgentProgress = {
  orchestrator?: 'pending' | 'running' | 'done' | 'failed';
  research?: 'pending' | 'running' | 'done' | 'failed';
  businessPlan?: 'pending' | 'running' | 'done' | 'failed';
  mvpBuilder?: 'pending' | 'running' | 'done' | 'failed';
} | null;

export interface Session {
  id: string;
  title: string;
  idea: string;
  status: SessionStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  agentProgress?: SessionAgentProgress;
}

export interface SessionInput {
  /** @minLength 10 */
  idea: string;
}

export type SessionDetailStatus = typeof SessionDetailStatus[keyof typeof SessionDetailStatus];


export const SessionDetailStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type SessionDetailAgentProgress = {
  orchestrator?: string;
  research?: string;
  businessPlan?: string;
  mvpBuilder?: string;
} | null;

/**
 * @nullable
 */
export type SessionDetailOrchestratorResult = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type SessionDetailResearchResult = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type SessionDetailBusinessPlanResult = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type SessionDetailMvpResult = { [key: string]: unknown } | null;

export interface SessionDetail {
  id: string;
  title: string;
  idea: string;
  status: SessionDetailStatus;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  agentProgress?: SessionDetailAgentProgress;
  /** @nullable */
  orchestratorResult?: SessionDetailOrchestratorResult;
  /** @nullable */
  researchResult?: SessionDetailResearchResult;
  /** @nullable */
  businessPlanResult?: SessionDetailBusinessPlanResult;
  /** @nullable */
  mvpResult?: SessionDetailMvpResult;
  /** @nullable */
  gitlabUrl?: string | null;
  /** @nullable */
  memoryCount?: number | null;
}

export type AgentStatusStatus = typeof AgentStatusStatus[keyof typeof AgentStatusStatus];


export const AgentStatusStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * @nullable
 */
export type AgentStatusAgentProgress = { [key: string]: unknown } | null;

export interface AgentStatus {
  sessionId: string;
  status: AgentStatusStatus;
  /** @nullable */
  agentProgress?: AgentStatusAgentProgress;
  /** @nullable */
  error?: string | null;
}

export interface RecallQuery {
  /** @minLength 1 */
  query: string;
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  limit?: number | null;
}

/**
 * @nullable
 */
export type MemoryRecordMetadata = { [key: string]: unknown } | null;

export interface MemoryRecord {
  id: string;
  sessionId: string;
  type: string;
  content: string;
  /** @nullable */
  metadata?: MemoryRecordMetadata;
  /** @nullable */
  score?: number | null;
  createdAt: string;
}

export interface DashboardStats {
  totalSessions: number;
  completedSessions: number;
  totalMemories: number;
  gitlabRepos?: number;
  recentSessions: Session[];
}

export interface DeleteResult {
  success: boolean;
  message?: string;
}

export type ListMemoriesParams = {
sessionId?: string;
type?: string;
limit?: number;
};