# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FounderAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory/recall:
    post:
      operationId: recallMemory
      tags: [memory]
      summary: Semantic vector search over all stored memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecallQuery"
      responses:
        "200":
          description: Recalled memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryRecord"
  /memory/list:
    get:
      operationId: listMemories
      tags: [memory]
      summary: List recent memories with optional session filter
      parameters:
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryRecord"
...
components:
  schemas:
...
    RecallQuery:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        sessionId:
          type: ["string", "null"]
        limit:
          type: ["integer", "null"]
    MemoryRecord:
      type: object
      required: [id, sessionId, type, content, createdAt]
      properties:
        id:
          type: string
        sessionId:
          type: string
        type:
          type: string
        content:
          type: string
        metadata:
          type: ["object", "null"]
          additionalProperties: true
        score:
          type: ["number", "null"]
        createdAt:
          type: string
...