# lib/api-zod/src/generated/types/memoryRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FounderAI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryRecordMetadata } from './memoryRecordMetadata';

export interface MemoryRecord {
  id: string;
  sessionId: string;
  type: string;
  content: string;
  /** @nullable */
  metadata?: MemoryRecordMetadata;
  /** @nullable */
  score?: number | null;
  createdAt: string;
}